Fixing querying of end-points from Grid.
Closes #2

It looks like the recent versions of Grid have
issues when handling POSTs for Grid APIs. This
is now cascading back to clients such as talk2grid
which rely on these end-points.

So until the Grid API end-points are fixed to handle
both GET and POST, altering “talk2grid” to use 
GET instead of POST.

Also upgraded testng to latest released version.